Make autotrain work
